test: move enterprise-only regression tests to *-ent spec files
The service graph toggle (Bug #11687) is gated behind
config.isEnterprise == 'true' in SearchBar.vue, and the
scheduled search jobs API (Bug #7354) returns 404 in OSS builds.
Both caused consistent failures in the OSS regression CI.

Move these tests into dedicated *-ent spec files that live in the
OSS repo but are only triggered by the enterprise CI workflow.

- logs-regression-ent.spec.js: Scheduled Search / search_jobs tests
- traces-regression-ent.spec.js: service graph side panel test
- Remove extracted tests from their original OSS spec files
